chore: Minor restructure and cleanup
