+++ build.rs
@@ -1,6 +1,10 @@
 fn main() {
   cxx_build::bridge("src/bridge.rs")
     .file("libcamera-bridge/core.cpp")
+    .file("libcamera-bridge/camera_manager.cpp")
+    .file("libcamera-bridge/camera.cpp")
+    .file("libcamera-bridge/frame_buffers.cpp")
+    .file("libcamera-bridge/config.cpp")
     .flag_if_supported("-std=c++17")
     .include("/usr/local/include/libcamera")
     .include("libcamera/build/include/libcamera")
